编程规范告警清除

+++ frameworks/native/pulseaudio/modules/hdi/hdi_sink.c
@@ -1549,7 +1549,7 @@ static void SinkRenderPrimaryProcess(pa_sink *si, size_t length, pa_memchunk *ch
     size_t memsetOutLen = sizeof(float) * DEFAULT_FRAMELEN * OUT_CHANNEL_NUM_MAX;
     memset_s(u->bufferAttr->tempBufIn, u->processSize, 0, memsetInLen);
     memset_s(u->bufferAttr->tempBufOut, u->processSize, 0, memsetOutLen);
-    int32_t bitSize = pa_sample_size_of_format(u->format);
+    int32_t bitSize = (int32_t)pa_sample_size_of_format(u->format);
     chunkIn->memblock = pa_memblock_new(si->core->mempool, length * IN_CHANNEL_NUM_MAX / DEFAULT_IN_CHANNEL_NUM);
     time_t currentTime = time(NULL);
     PrepareSpatializationFading(&u->spatializationFadingState, &u->spatializationFadingCount,
